requirements.txt
streamlit==1.10.0
langchain==0.0.1
langchain_openai==0.0.1
langchain_groq==0.0.1
langchain_community==0.0.1
langchain_astradb==0.0.1
unstructured==0.0.1
python-dotenv==0.19.2
uuid==1.30